nrf52840_hal

Module comp

HAL interface for the COMP peripheral.

The comparator (COMP) compares an input voltage (Vin) against a second input voltage (Vref). Vin can be derived from an analog input pin (AIN0-AIN7). Vref can be derived from multiple sources depending on the operation mode of the comparator.

Structs

Comp

A safe wrapper around the COMP peripheral.

Enums

CompResult

OperationMode

PowerMode

Transition

VRef

Traits

CompInputPin

Trait to represent analog input pins.

CompRefPin

Trait to represent analog ref pins.