nrf52840_hal::pac::nfct

Struct RegisterBlock

pub struct RegisterBlock {

    pub tasks_activate: Reg<TASKS_ACTIVATE_SPEC>,
    pub tasks_disable: Reg<TASKS_DISABLE_SPEC>,
    pub tasks_sense: Reg<TASKS_SENSE_SPEC>,
    pub tasks_starttx: Reg<TASKS_STARTTX_SPEC>,
    pub tasks_enablerxdata: Reg<TASKS_ENABLERXDATA_SPEC>,
    pub tasks_goidle: Reg<TASKS_GOIDLE_SPEC>,
    pub tasks_gosleep: Reg<TASKS_GOSLEEP_SPEC>,
    pub events_ready: Reg<EVENTS_READY_SPEC>,
    pub events_fielddetected: Reg<EVENTS_FIELDDETECTED_SPEC>,
    pub events_fieldlost: Reg<EVENTS_FIELDLOST_SPEC>,
    pub events_txframestart: Reg<EVENTS_TXFRAMESTART_SPEC>,
    pub events_txframeend: Reg<EVENTS_TXFRAMEEND_SPEC>,
    pub events_rxframestart: Reg<EVENTS_RXFRAMESTART_SPEC>,
    pub events_rxframeend: Reg<EVENTS_RXFRAMEEND_SPEC>,
    pub events_error: Reg<EVENTS_ERROR_SPEC>,
    pub events_rxerror: Reg<EVENTS_RXERROR_SPEC>,
    pub events_endrx: Reg<EVENTS_ENDRX_SPEC>,
    pub events_endtx: Reg<EVENTS_ENDTX_SPEC>,
    pub events_autocolresstarted: Reg<EVENTS_AUTOCOLRESSTARTED_SPEC>,
    pub events_collision: Reg<EVENTS_COLLISION_SPEC>,
    pub events_selected: Reg<EVENTS_SELECTED_SPEC>,
    pub events_started: Reg<EVENTS_STARTED_SPEC>,
    pub shorts: Reg<SHORTS_SPEC>,
    pub inten: Reg<INTEN_SPEC>,
    pub intenset: Reg<INTENSET_SPEC>,
    pub intenclr: Reg<INTENCLR_SPEC>,
    pub errorstatus: Reg<ERRORSTATUS_SPEC>,
    pub framestatus: FRAMESTATUS,
    pub nfctagstate: Reg<NFCTAGSTATE_SPEC>,
    pub sleepstate: Reg<SLEEPSTATE_SPEC>,
    pub fieldpresent: Reg<FIELDPRESENT_SPEC>,
    pub framedelaymin: Reg<FRAMEDELAYMIN_SPEC>,
    pub framedelaymax: Reg<FRAMEDELAYMAX_SPEC>,
    pub framedelaymode: Reg<FRAMEDELAYMODE_SPEC>,
    pub packetptr: Reg<PACKETPTR_SPEC>,
    pub maxlen: Reg<MAXLEN_SPEC>,
    pub txd: TXD,
    pub rxd: RXD,
    pub nfcid1_last: Reg<NFCID1_LAST_SPEC>,
    pub nfcid1_2nd_last: Reg<NFCID1_2ND_LAST_SPEC>,
    pub nfcid1_3rd_last: Reg<NFCID1_3RD_LAST_SPEC>,
    pub autocolresconfig: Reg<AUTOCOLRESCONFIG_SPEC>,
    pub sensres: Reg<SENSRES_SPEC>,
    pub selres: Reg<SELRES_SPEC>,
    /* private fields */

}

Register block

Fields
tasks_activate: Reg<TASKS_ACTIVATE_SPEC>

0x00 - Activate NFCT peripheral for incoming and outgoing frames, change state to activated

tasks_disable: Reg<TASKS_DISABLE_SPEC>

0x04 - Disable NFCT peripheral

tasks_sense: Reg<TASKS_SENSE_SPEC>

0x08 - Enable NFC sense field mode, change state to sense mode

tasks_starttx: Reg<TASKS_STARTTX_SPEC>

0x0c - Start transmission of an outgoing frame, change state to transmit

tasks_enablerxdata: Reg<TASKS_ENABLERXDATA_SPEC>

0x1c - Initializes the EasyDMA for receive.

tasks_goidle: Reg<TASKS_GOIDLE_SPEC>

0x24 - Force state machine to IDLE state

tasks_gosleep: Reg<TASKS_GOSLEEP_SPEC>

0x28 - Force state machine to SLEEP_A state

events_ready: Reg<EVENTS_READY_SPEC>

0x100 - The NFCT peripheral is ready to receive and send frames

events_fielddetected: Reg<EVENTS_FIELDDETECTED_SPEC>

0x104 - Remote NFC field detected

events_fieldlost: Reg<EVENTS_FIELDLOST_SPEC>

0x108 - Remote NFC field lost

events_txframestart: Reg<EVENTS_TXFRAMESTART_SPEC>

0x10c - Marks the start of the first symbol of a transmitted frame

events_txframeend: Reg<EVENTS_TXFRAMEEND_SPEC>

0x110 - Marks the end of the last transmitted on-air symbol of a frame

events_rxframestart: Reg<EVENTS_RXFRAMESTART_SPEC>

0x114 - Marks the end of the first symbol of a received frame

events_rxframeend: Reg<EVENTS_RXFRAMEEND_SPEC>

0x118 - Received data has been checked (CRC, parity) and transferred to RAM, and EasyDMA has ended accessing the RX buffer

events_error: Reg<EVENTS_ERROR_SPEC>

0x11c - NFC error reported. The ERRORSTATUS register contains details on the source of the error.

events_rxerror: Reg<EVENTS_RXERROR_SPEC>

0x128 - NFC RX frame error reported. The FRAMESTATUS.RX register contains details on the source of the error.

events_endrx: Reg<EVENTS_ENDRX_SPEC>

0x12c - RX buffer (as defined by PACKETPTR and MAXLEN) in Data RAM full.

events_endtx: Reg<EVENTS_ENDTX_SPEC>

0x130 - Transmission of data in RAM has ended, and EasyDMA has ended accessing the TX buffer

events_autocolresstarted: Reg<EVENTS_AUTOCOLRESSTARTED_SPEC>

0x138 - Auto collision resolution process has started

events_collision: Reg<EVENTS_COLLISION_SPEC>

0x148 - NFC auto collision resolution error reported.

events_selected: Reg<EVENTS_SELECTED_SPEC>

0x14c - NFC auto collision resolution successfully completed

events_started: Reg<EVENTS_STARTED_SPEC>

0x150 - EasyDMA is ready to receive or send frames.

shorts: Reg<SHORTS_SPEC>

0x200 - Shortcut register

inten: Reg<INTEN_SPEC>

0x300 - Enable or disable interrupt

intenset: Reg<INTENSET_SPEC>

0x304 - Enable interrupt

intenclr: Reg<INTENCLR_SPEC>

0x308 - Disable interrupt

errorstatus: Reg<ERRORSTATUS_SPEC>

0x404 - NFC Error Status register

framestatus: FRAMESTATUS

0x40c - Unspecified

nfctagstate: Reg<NFCTAGSTATE_SPEC>

0x410 - NfcTag state register

sleepstate: Reg<SLEEPSTATE_SPEC>

0x420 - Sleep state during automatic collision resolution

fieldpresent: Reg<FIELDPRESENT_SPEC>

0x43c - Indicates the presence or not of a valid field

framedelaymin: Reg<FRAMEDELAYMIN_SPEC>

0x504 - Minimum frame delay

framedelaymax: Reg<FRAMEDELAYMAX_SPEC>

0x508 - Maximum frame delay

framedelaymode: Reg<FRAMEDELAYMODE_SPEC>

0x50c - Configuration register for the Frame Delay Timer

packetptr: Reg<PACKETPTR_SPEC>

0x510 - Packet pointer for TXD and RXD data storage in Data RAM

maxlen: Reg<MAXLEN_SPEC>

0x514 - Size of the RAM buffer allocated to TXD and RXD data storage each

txd: TXD

0x518..0x520 - Unspecified

rxd: RXD

0x520..0x528 - Unspecified

nfcid1_last: Reg<NFCID1_LAST_SPEC>

0x590 - Last NFCID1 part (4, 7 or 10 bytes ID)

nfcid1_2nd_last: Reg<NFCID1_2ND_LAST_SPEC>

0x594 - Second last NFCID1 part (7 or 10 bytes ID)

nfcid1_3rd_last: Reg<NFCID1_3RD_LAST_SPEC>

0x598 - Third last NFCID1 part (10 bytes ID)

autocolresconfig: Reg<AUTOCOLRESCONFIG_SPEC>

0x59c - Controls the auto collision resolution function. This setting must be done before the NFCT peripheral is enabled.

sensres: Reg<SENSRES_SPEC>

0x5a0 - NFC-A SENS_RES auto-response settings

selres: Reg<SELRES_SPEC>

0x5a4 - NFC-A SEL_RES auto-response settings