nrf52840_hal::pac::radio

Struct RegisterBlock

pub struct RegisterBlock {

    pub tasks_txen: Reg<TASKS_TXEN_SPEC>,
    pub tasks_rxen: Reg<TASKS_RXEN_SPEC>,
    pub tasks_start: Reg<TASKS_START_SPEC>,
    pub tasks_stop: Reg<TASKS_STOP_SPEC>,
    pub tasks_disable: Reg<TASKS_DISABLE_SPEC>,
    pub tasks_rssistart: Reg<TASKS_RSSISTART_SPEC>,
    pub tasks_rssistop: Reg<TASKS_RSSISTOP_SPEC>,
    pub tasks_bcstart: Reg<TASKS_BCSTART_SPEC>,
    pub tasks_bcstop: Reg<TASKS_BCSTOP_SPEC>,
    pub tasks_edstart: Reg<TASKS_EDSTART_SPEC>,
    pub tasks_edstop: Reg<TASKS_EDSTOP_SPEC>,
    pub tasks_ccastart: Reg<TASKS_CCASTART_SPEC>,
    pub tasks_ccastop: Reg<TASKS_CCASTOP_SPEC>,
    pub events_ready: Reg<EVENTS_READY_SPEC>,
    pub events_address: Reg<EVENTS_ADDRESS_SPEC>,
    pub events_payload: Reg<EVENTS_PAYLOAD_SPEC>,
    pub events_end: Reg<EVENTS_END_SPEC>,
    pub events_disabled: Reg<EVENTS_DISABLED_SPEC>,
    pub events_devmatch: Reg<EVENTS_DEVMATCH_SPEC>,
    pub events_devmiss: Reg<EVENTS_DEVMISS_SPEC>,
    pub events_rssiend: Reg<EVENTS_RSSIEND_SPEC>,
    pub events_bcmatch: Reg<EVENTS_BCMATCH_SPEC>,
    pub events_crcok: Reg<EVENTS_CRCOK_SPEC>,
    pub events_crcerror: Reg<EVENTS_CRCERROR_SPEC>,
    pub events_framestart: Reg<EVENTS_FRAMESTART_SPEC>,
    pub events_edend: Reg<EVENTS_EDEND_SPEC>,
    pub events_edstopped: Reg<EVENTS_EDSTOPPED_SPEC>,
    pub events_ccaidle: Reg<EVENTS_CCAIDLE_SPEC>,
    pub events_ccabusy: Reg<EVENTS_CCABUSY_SPEC>,
    pub events_ccastopped: Reg<EVENTS_CCASTOPPED_SPEC>,
    pub events_rateboost: Reg<EVENTS_RATEBOOST_SPEC>,
    pub events_txready: Reg<EVENTS_TXREADY_SPEC>,
    pub events_rxready: Reg<EVENTS_RXREADY_SPEC>,
    pub events_mhrmatch: Reg<EVENTS_MHRMATCH_SPEC>,
    pub events_phyend: Reg<EVENTS_PHYEND_SPEC>,
    pub shorts: Reg<SHORTS_SPEC>,
    pub intenset: Reg<INTENSET_SPEC>,
    pub intenclr: Reg<INTENCLR_SPEC>,
    pub crcstatus: Reg<CRCSTATUS_SPEC>,
    pub rxmatch: Reg<RXMATCH_SPEC>,
    pub rxcrc: Reg<RXCRC_SPEC>,
    pub dai: Reg<DAI_SPEC>,
    pub pdustat: Reg<PDUSTAT_SPEC>,
    pub packetptr: Reg<PACKETPTR_SPEC>,
    pub frequency: Reg<FREQUENCY_SPEC>,
    pub txpower: Reg<TXPOWER_SPEC>,
    pub mode: Reg<MODE_SPEC>,
    pub pcnf0: Reg<PCNF0_SPEC>,
    pub pcnf1: Reg<PCNF1_SPEC>,
    pub base0: Reg<BASE0_SPEC>,
    pub base1: Reg<BASE1_SPEC>,
    pub prefix0: Reg<PREFIX0_SPEC>,
    pub prefix1: Reg<PREFIX1_SPEC>,
    pub txaddress: Reg<TXADDRESS_SPEC>,
    pub rxaddresses: Reg<RXADDRESSES_SPEC>,
    pub crccnf: Reg<CRCCNF_SPEC>,
    pub crcpoly: Reg<CRCPOLY_SPEC>,
    pub crcinit: Reg<CRCINIT_SPEC>,
    pub tifs: Reg<TIFS_SPEC>,
    pub rssisample: Reg<RSSISAMPLE_SPEC>,
    pub state: Reg<STATE_SPEC>,
    pub datawhiteiv: Reg<DATAWHITEIV_SPEC>,
    pub bcc: Reg<BCC_SPEC>,
    pub dab: [Reg<DAB_SPEC>; 8],
    pub dap: [Reg<DAP_SPEC>; 8],
    pub dacnf: Reg<DACNF_SPEC>,
    pub mhrmatchconf: Reg<MHRMATCHCONF_SPEC>,
    pub mhrmatchmas: Reg<MHRMATCHMAS_SPEC>,
    pub modecnf0: Reg<MODECNF0_SPEC>,
    pub sfd: Reg<SFD_SPEC>,
    pub edcnt: Reg<EDCNT_SPEC>,
    pub edsample: Reg<EDSAMPLE_SPEC>,
    pub ccactrl: Reg<CCACTRL_SPEC>,
    pub power: Reg<POWER_SPEC>,
    /* private fields */

}

Register block

Fields
tasks_txen: Reg<TASKS_TXEN_SPEC>

0x00 - Enable RADIO in TX mode

tasks_rxen: Reg<TASKS_RXEN_SPEC>

0x04 - Enable RADIO in RX mode

tasks_start: Reg<TASKS_START_SPEC>

0x08 - Start RADIO

tasks_stop: Reg<TASKS_STOP_SPEC>

0x0c - Stop RADIO

tasks_disable: Reg<TASKS_DISABLE_SPEC>

0x10 - Disable RADIO

tasks_rssistart: Reg<TASKS_RSSISTART_SPEC>

0x14 - Start the RSSI and take one single sample of the receive signal strength

tasks_rssistop: Reg<TASKS_RSSISTOP_SPEC>

0x18 - Stop the RSSI measurement

tasks_bcstart: Reg<TASKS_BCSTART_SPEC>

0x1c - Start the bit counter

tasks_bcstop: Reg<TASKS_BCSTOP_SPEC>

0x20 - Stop the bit counter

tasks_edstart: Reg<TASKS_EDSTART_SPEC>

0x24 - Start the energy detect measurement used in IEEE 802.15.4 mode

tasks_edstop: Reg<TASKS_EDSTOP_SPEC>

0x28 - Stop the energy detect measurement

tasks_ccastart: Reg<TASKS_CCASTART_SPEC>

0x2c - Start the clear channel assessment used in IEEE 802.15.4 mode

tasks_ccastop: Reg<TASKS_CCASTOP_SPEC>

0x30 - Stop the clear channel assessment

events_ready: Reg<EVENTS_READY_SPEC>

0x100 - RADIO has ramped up and is ready to be started

events_address: Reg<EVENTS_ADDRESS_SPEC>

0x104 - Address sent or received

events_payload: Reg<EVENTS_PAYLOAD_SPEC>

0x108 - Packet payload sent or received

events_end: Reg<EVENTS_END_SPEC>

0x10c - Packet sent or received

events_disabled: Reg<EVENTS_DISABLED_SPEC>

0x110 - RADIO has been disabled

events_devmatch: Reg<EVENTS_DEVMATCH_SPEC>

0x114 - A device address match occurred on the last received packet

events_devmiss: Reg<EVENTS_DEVMISS_SPEC>

0x118 - No device address match occurred on the last received packet

events_rssiend: Reg<EVENTS_RSSIEND_SPEC>

0x11c - Sampling of receive signal strength complete

events_bcmatch: Reg<EVENTS_BCMATCH_SPEC>

0x128 - Bit counter reached bit count value

events_crcok: Reg<EVENTS_CRCOK_SPEC>

0x130 - Packet received with CRC ok

events_crcerror: Reg<EVENTS_CRCERROR_SPEC>

0x134 - Packet received with CRC error

events_framestart: Reg<EVENTS_FRAMESTART_SPEC>

0x138 - IEEE 802.15.4 length field received

events_edend: Reg<EVENTS_EDEND_SPEC>

0x13c - Sampling of energy detection complete. A new ED sample is ready for readout from the RADIO.EDSAMPLE register.

events_edstopped: Reg<EVENTS_EDSTOPPED_SPEC>

0x140 - The sampling of energy detection has stopped

events_ccaidle: Reg<EVENTS_CCAIDLE_SPEC>

0x144 - Wireless medium in idle - clear to send

events_ccabusy: Reg<EVENTS_CCABUSY_SPEC>

0x148 - Wireless medium busy - do not send

events_ccastopped: Reg<EVENTS_CCASTOPPED_SPEC>

0x14c - The CCA has stopped

events_rateboost: Reg<EVENTS_RATEBOOST_SPEC>

0x150 - Ble_LR CI field received, receive mode is changed from Ble_LR125Kbit to Ble_LR500Kbit.

events_txready: Reg<EVENTS_TXREADY_SPEC>

0x154 - RADIO has ramped up and is ready to be started TX path

events_rxready: Reg<EVENTS_RXREADY_SPEC>

0x158 - RADIO has ramped up and is ready to be started RX path

events_mhrmatch: Reg<EVENTS_MHRMATCH_SPEC>

0x15c - MAC header match found

events_phyend: Reg<EVENTS_PHYEND_SPEC>

0x16c - Generated in Ble_LR125Kbit, Ble_LR500Kbit and BleIeee802154_250Kbit modes when last bit is sent on air.

shorts: Reg<SHORTS_SPEC>

0x200 - Shortcut register

intenset: Reg<INTENSET_SPEC>

0x304 - Enable interrupt

intenclr: Reg<INTENCLR_SPEC>

0x308 - Disable interrupt

crcstatus: Reg<CRCSTATUS_SPEC>

0x400 - CRC status

rxmatch: Reg<RXMATCH_SPEC>

0x408 - Received address

rxcrc: Reg<RXCRC_SPEC>

0x40c - CRC field of previously received packet

dai: Reg<DAI_SPEC>

0x410 - Device address match index

pdustat: Reg<PDUSTAT_SPEC>

0x414 - Payload status

packetptr: Reg<PACKETPTR_SPEC>

0x504 - Packet pointer

frequency: Reg<FREQUENCY_SPEC>

0x508 - Frequency

txpower: Reg<TXPOWER_SPEC>

0x50c - Output power

mode: Reg<MODE_SPEC>

0x510 - Data rate and modulation

pcnf0: Reg<PCNF0_SPEC>

0x514 - Packet configuration register 0

pcnf1: Reg<PCNF1_SPEC>

0x518 - Packet configuration register 1

base0: Reg<BASE0_SPEC>

0x51c - Base address 0

base1: Reg<BASE1_SPEC>

0x520 - Base address 1

prefix0: Reg<PREFIX0_SPEC>

0x524 - Prefixes bytes for logical addresses 0-3

prefix1: Reg<PREFIX1_SPEC>

0x528 - Prefixes bytes for logical addresses 4-7

txaddress: Reg<TXADDRESS_SPEC>

0x52c - Transmit address select

rxaddresses: Reg<RXADDRESSES_SPEC>

0x530 - Receive address select

crccnf: Reg<CRCCNF_SPEC>

0x534 - CRC configuration

crcpoly: Reg<CRCPOLY_SPEC>

0x538 - CRC polynomial

crcinit: Reg<CRCINIT_SPEC>

0x53c - CRC initial value

tifs: Reg<TIFS_SPEC>

0x544 - Interframe spacing in us

rssisample: Reg<RSSISAMPLE_SPEC>

0x548 - RSSI sample

state: Reg<STATE_SPEC>

0x550 - Current radio state

datawhiteiv: Reg<DATAWHITEIV_SPEC>

0x554 - Data whitening initial value

bcc: Reg<BCC_SPEC>

0x560 - Bit counter compare

dab: [Reg<DAB_SPEC>; 8]

0x600..0x620 - Description collection[n]: Device address base segment n

dap: [Reg<DAP_SPEC>; 8]

0x620..0x640 - Description collection[n]: Device address prefix n

dacnf: Reg<DACNF_SPEC>

0x640 - Device address match configuration

mhrmatchconf: Reg<MHRMATCHCONF_SPEC>

0x644 - Search pattern configuration

mhrmatchmas: Reg<MHRMATCHMAS_SPEC>

0x648 - Pattern mask

modecnf0: Reg<MODECNF0_SPEC>

0x650 - Radio mode configuration register 0

sfd: Reg<SFD_SPEC>

0x660 - IEEE 802.15.4 start of frame delimiter

edcnt: Reg<EDCNT_SPEC>

0x664 - IEEE 802.15.4 energy detect loop count

edsample: Reg<EDSAMPLE_SPEC>

0x668 - IEEE 802.15.4 energy detect level

ccactrl: Reg<CCACTRL_SPEC>

0x66c - IEEE 802.15.4 clear channel assessment control

power: Reg<POWER_SPEC>

0xffc - Peripheral power control