nrf52840_hal

Module ppi

HAL interface for the PPI peripheral.

The Programmable Peripheral Interconnect interface allows for an autonomous interoperability between peripherals through their events and tasks. There are fixed PPI channels and fully configurable ones, fixed channels can only connect specific events to specific tasks. For fully configurable channels, it is possible to choose, via software, the event and the task that it will triggered by the event.

On nRF52 devices, there is also a fork task endpoint, where the user can configure one more task to be triggered by the same event, even fixed PPI channels have a configurable fork task.

Structs

Chg0

Channel groups.

Chg1

Channel groups.

Chg2

Channel groups.

Chg3

Channel groups.

Chg4

Channel groups.

Chg5

Channel groups.

EventAddr

Parts

Type that abstracts all the PPI channels.

Ppi0

Fully configurable PPI Channel.

Ppi1

Fully configurable PPI Channel.

Ppi2

Fully configurable PPI Channel.

Ppi3

Fully configurable PPI Channel.

Ppi4

Fully configurable PPI Channel.

Ppi5

Fully configurable PPI Channel.

Ppi6

Fully configurable PPI Channel.

Ppi7

Fully configurable PPI Channel.

Ppi8

Fully configurable PPI Channel.

Ppi9

Fully configurable PPI Channel.

Ppi10

Fully configurable PPI Channel.

Ppi11

Fully configurable PPI Channel.

Ppi12

Fully configurable PPI Channel.

Ppi13

Fully configurable PPI Channel.

Ppi14

Fully configurable PPI Channel.

Ppi15

Fully configurable PPI Channel.

Ppi16

Fully configurable PPI Channel.

Ppi17

Fully configurable PPI Channel.

Ppi18

Fully configurable PPI Channel.

Ppi19

Fully configurable PPI Channel.

Ppi20

Fixed PPI channel.

Ppi21

Fixed PPI channel.

Ppi22

Fixed PPI channel.

Ppi23

Fixed PPI channel.

Ppi24

Fixed PPI channel.

Ppi25

Fixed PPI channel.

Ppi26

Fixed PPI channel.

Ppi27

Fixed PPI channel.

Ppi28

Fixed PPI channel.

Ppi29

Fixed PPI channel.

Ppi30

Fixed PPI channel.

Ppi31

Fixed PPI channel.

TaskAddr

Traits

ConfigurablePpi

Traits that extends the Ppi trait, marking a channel as fully configurable.

Ppi

Trait to represent a Programmable Peripheral Interconnect channel.

PpiChannelGroup

Trait for a PPI channel group.