nrf52840_hal::ppi

Trait ConfigurablePpi

pub trait ConfigurablePpi: Ppi {
    // Required methods
    fn set_task_endpoint<T>(&mut self, task: &T)
       where T: Task;
    fn set_event_endpoint<E>(&mut self, event: &E)
       where E: Event;
}

Traits that extends the Ppi trait, marking a channel as fully configurable.

Required Methods

fn set_task_endpoint<T>(&mut self, task: &T)
where T: Task,

Sets the task that must be triggered when the configured event occurs. The user must provide a reference to the task.

fn set_event_endpoint<E>(&mut self, event: &E)
where E: Event,

Sets the event that will trigger the chosen task(s). The user must provide a reference to the event.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.