nrf52840_hal

Module uarte

HAL interface to the UARTE peripheral.

See product specification:

nrf52832: Section 35

nrf52840: Section 6.34

Structs

Pins

Uarte

Interface to a UARTE instance.

UarteRx

Interface for the RX part of a UART instance that can be used independently of the TX part.

UarteTx

Interface for the TX part of a UART instance that can be used independently of the RX part.

Enums

Baudrate

Baud rate

Error

Parity

Parity

Traits

Instance