nrf52840_hal

Module wdt

HAL interface to the WDT peripheral.

This HAL implements a basic watchdog timer with 1..=8 handles. Once the watchdog has been started, it cannot be stopped.

Modules

count

Type states representing the number of requested handles.

handles

Type states representing individual watchdog handles.

Structs

Active

A type state representing a watchdog that has been started and cannot be stopped.

Inactive

A type state representing a watchdog that has not been started.

Parts

A structure containing the active watchdog and all requested Watchdog handles.

Watchdog

An interface to the Watchdog.

WatchdogHandle

An interface to feed the Watchdog.